Add record type PatternCharMaps. Add folder for record types. Move record types there.
