+++ CarRental/Models/Car/ServiceManager.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
-using CarRental.Models.Car;
 using CarRental.Enumerables;
 using CarRental.Interfaces;
+using CarRental.Models.Car.RecordTypes;
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
 // HOMEWORK: "ДЗ 3. Methods, properties"
