+++ CarRental/Models/Car/ServiceManagerSupplements.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using CarRental.Models.Car.RecordTypes;
 using static System.Runtime.InteropServices.JavaScript.JSType;
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
